Update default AWS region in BedrockClientSettings to us-west-2
