+++ prompt/prompt-executor/prompt-executor-clients/prompt-executor-bedrock-client/src/jvmTest/kotlin/ai/koog/prompt/executor/clients/bedrock/BedrockLLMClientTest.kt
@@ -402,7 +402,6 @@ class BedrockLLMClientTest {
 
     @Test
     fun testToolChoiceConfiguration() {
-
         // Test different tool choice configurations
         val autoPrompt = Prompt.build("test", params = LLMParams(toolChoice = LLMParams.ToolChoice.Auto)) {
             user("Search for something")
